[mapleall]Internal commit msg:     !1154 ebo extend opt for 'and 0xff 0xffff' and 'back to back and'     !1156 Handle negative offset in globalopt. no ssa opt for phys reg.

+++ src/mapleall/maple_be/src/cg/reaching.cpp
@@ -778,7 +778,7 @@ void ReachingDefinition::AnalysisStart() {
   if (!cgFunc->GetFirstBB()) {
     return;
   }
-  stackSize = GetStackSize();
+  stackSize = static_cast<uint32>(GetStackSize());
   Initialize();
   /* Build in/out for function body first. (Except cleanup bb) */
   BuildInOutForFuncBody();
